Rework $resparent variable to work with arrays and update queries
In order to accommodate the discovery help pages the research guides A-Z needs to retrieve guides from
the children of two pages - in order to accommodate this the query now takes an array of pages whose
children to include.

The page IDs are declared in a constant in the envisronment constants file.