Fix incorrect <ul> structure on research guide results
This is in relation to an accessibility issue where all research guide results were under one ul with h3's inside to separate results. This has been refactored so that each h3 is outside of a ul, and there are multiple ul elements on the page now for greater separation.