Remove script which duplicates HTML for sub-menu
When this script was introduced there were no 'home' links included in the
mega menu HTML. This script therefore created them so that 'mobile' users would
be able to access them when the main sections were converted to accordion headers.

It seems that since this time the 'home' links have been introduced to the HTML
which is resulting in the links being duplicated. To avoid this problem it seems
most sensible to remove the script.